Touch repomd.xml files to update the timestamp
We check the repomd.xml file of each mirror to see if it's up-to-date
(modified in the last 24 hours) before adding a mirror to the mirrorlist.
However, if the RPMs haven't changed, then "createrepo --update" will not
modify repomd.xml (including the timestamp), meaning the up-to-dateness
check will fail for all mirrors (including the fallback mirror (the repo
server itself)).

This only affects the {23,24}-{contrib,empty,internal} tags -- these are
the tags that do not have condor repos defined for them.
The directories that the condor repos get rsynced to always have timestamps
matching the rsync, which I guess is enough to change the timestamps in the
repo metadata despite there being no changes to any of the files.